Optimize Sylvester Stallone

+++ catscripts/Entertainment/Actors_and_actresses/Sylvester_Stallone/SylvesterStallone.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-export KEYWORDS_ACTOR_SYLVESTERSTALLONE="Sylvester(| )Stallone"
+KEYWORDS_ACTOR_SYLVESTERSTALLONE="Sylvester(| )Stallone"
 
-egrep -i "$KEYWORDS_ACTOR_SYLVESTERSTALLONE" newpages.txt >> SylvesterStallone.txt
-
-SYLVESTERSTALLONE=`stat --print=%s SylvesterStallone.txt`
-
-if [ $SYLVESTERSTALLONE -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="SylvesterStallone.txt"
-  export CATNAME="Sylvester Stallone"
-  $CATEGORIZE
-fi
 
-rm SylvesterStallone.txt
+  SYLVESTERSTALLONE=`egrep -i "$KEYWORDS_ACTOR_SYLVESTERSTALLONE" newpages.txt`
+
+  if [ $SYLVESTERSTALLONE -ne 0 ];
+  then
+    egrep -i "$KEYWORDS_ACTOR_SYLVESTERSTALLONE" newpages.txt > SylvesterStallone.txt
+    export CATFILE="SylvesterStallone.txt"
+    export CATNAME="Sylvester Stallone"
+    $CATEGORIZE
+    rm SylvesterStallone.txt
+    unset SYLVESTERSTALLONE
+  fi
+
+fi
